@@ -2,13 +2,13 @@ package i118Utils
 
 import (
 	"encoding/json"
+	"github.com/easysoft/zendata"
 	constant "github.com/easysoft/zendata/internal/pkg/const"
 	"github.com/easysoft/zendata/pkg/utils/common"
 	"github.com/easysoft/zendata/pkg/utils/vari"
-	"github.com/easysoft/zendata/res"
 	"golang.org/x/text/language"
 	"golang.org/x/text/message"
-	"io/ioutil"
+	"os"
 	"path/filepath"
 )
 
@@ -27,10 +27,10 @@ func InitI118(lang string) {
 	}
 
 	if isRelease {
-		data, _ := res.Asset(langRes)
+		data, _ := zd.ReadResData(langRes)
 		InitResFromAsset(data)
 	} else {
-		InitRes(filepath.Join(vari.ZdPath, langRes))
+		InitRes(filepath.Join(vari.DevDir, langRes))
 	}
 
 	if lang == "zh" {
@@ -58,7 +58,7 @@ func Check(e error) {
 	}
 }
 func ReadI18nJson(file string) string {
-	b, err := ioutil.ReadFile(file)
+	b, err := os.ReadFile(file)
 	Check(err)
 	str := string(b)
 	return str
