@@ -4,9 +4,9 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
+	"github.com/easysoft/zendata/cmd/test/proto/defaults"
+	"github.com/easysoft/zendata/cmd/test/proto/dist"
 	"github.com/easysoft/zendata/internal/pkg/model"
-	"github.com/easysoft/zendata/internal/pkg/test/proto/defaults"
-	"github.com/easysoft/zendata/internal/pkg/test/proto/dist"
 	fileUtils "github.com/easysoft/zendata/pkg/utils/file"
 	httpUtils "github.com/easysoft/zendata/pkg/utils/http"
 	stringUtils "github.com/easysoft/zendata/pkg/utils/string"
