@@ -1,11 +1,11 @@
-package helper
+package genHelper
 
 import (
-	"github.com/easysoft/zendata/internal/pkg/model"
+	"github.com/easysoft/zendata/internal/pkg/domain"
 	"strings"
 )
 
-func SelectExcelWithExpr(field model.DefField) bool {
+func IsSelectExcelWithExpr(field domain.DefField) bool {
 	if strings.Index(field.Select, "${") > -1 || strings.Index(field.Where, "${") > -1 {
 		return true
 	} else {
