@@ -21,7 +21,6 @@ func GenerateFromProtobuf(protoFile string) (content, pth string) {
 	outputDir := generateCls(protoFile)
 
 	convertFile := generateConverter(outputDir)
-
 	content, pth = generateBinData(convertFile)
 
 	return
@@ -32,7 +31,7 @@ func generateBinData(convertFile string) (content, pth string) {
 
 	phpExeFile := "php"
 	if commonUtils.IsWin() { // use build-in php runtime
-		phpExeFile = filepath.Join(vari.ZdPath, "runtime", "php", "php7", "php.exe")
+		phpExeFile = filepath.Join(vari.ZdDir, "runtime", "php", "php7", "php.exe")
 	}
 	cmdStr := phpExeFile + " convert.php"
 	out, _ := shellUtils.ExecInDir(cmdStr, dir)
@@ -46,7 +45,7 @@ func generateBinData(convertFile string) (content, pth string) {
 }
 
 func generateConverter(dir string) (pth string) {
-	srcFile := filepath.Join(vari.ZdPath, "runtime", "protobuf", "convert.php")
+	srcFile := filepath.Join(vari.ZdDir, "runtime", "protobuf", "convert.php")
 	pth = filepath.Join(dir, "convert.php")
 
 	content := fileUtils.ReadFile(srcFile)
@@ -69,7 +68,7 @@ func generateCls(protoFile string) (ret string) {
 		execFile += ".exe"
 	}
 
-	execFile = filepath.Join(vari.ZdPath, "runtime", "protobuf", "bin", platform, execFile)
+	execFile = filepath.Join(vari.ZdDir, "runtime", "protobuf", "bin", platform, execFile)
 
 	cmdStr := fmt.Sprintf("%s --php_out=%s %s", execFile, outputDir, protoFile)
 	shellUtils.Exec(cmdStr)
