@@ -1,8 +1,9 @@
 package gen
 
 import (
-	constant "github.com/easysoft/zendata/internal/pkg/const"
-	"github.com/easysoft/zendata/internal/pkg/model"
+	consts "github.com/easysoft/zendata/internal/pkg/const"
+	"github.com/easysoft/zendata/internal/pkg/domain"
+	"github.com/easysoft/zendata/internal/pkg/helper"
 	fileUtils "github.com/easysoft/zendata/pkg/utils/file"
 	i118Utils "github.com/easysoft/zendata/pkg/utils/i118"
 	logUtils "github.com/easysoft/zendata/pkg/utils/log"
@@ -16,22 +17,22 @@ import (
 func LoadResDef(fieldsToExport []string) (res map[string]map[string][]string) {
 	res = map[string]map[string][]string{}
 
-	for index, field := range vari.Def.Fields {
+	for index, field := range vari.GlobalVars.DefData.Fields {
 		if !stringUtils.StrInArr(field.Field, fieldsToExport) {
 			continue
 		}
 
 		if (field.Use != "" || field.Select != "") && field.From == "" {
-			field.From = vari.Def.From
-			vari.Def.Fields[index].From = vari.Def.From
+			field.From = vari.GlobalVars.DefData.From
+			vari.GlobalVars.DefData.Fields[index].From = vari.GlobalVars.DefData.From
 		}
 		loadResForFieldRecursive(&field, &res)
 	}
 	return
 }
 
-func loadResForFieldRecursive(field *model.DefField, res *map[string]map[string][]string) {
-	if len(field.Fields) > 0 { // sub fields
+func loadResForFieldRecursive(field *domain.DefField, res *map[string]map[string][]string) {
+	if len(field.Fields) > 0 { // child fields
 		for _, child := range field.Fields {
 			if child.Use != "" && child.From == "" {
 				child.From = field.From
@@ -50,21 +51,21 @@ func loadResForFieldRecursive(field *model.DefField, res *map[string]map[string]
 			loadResForFieldRecursive(&child, res)
 		}
 
-	} else if field.From != "" && field.Type != constant.FieldTypeArticle { // from a res
+	} else if field.From != "" && field.Type != consts.FieldTypeArticle { // from a res
 		var valueMap map[string][]string
 		resFile, resType, sheet := fileUtils.GetResProp(field.From, field.FileDir) // relate to current file
 		valueMap, _ = getResValue(resFile, resType, sheet, field)
 
-		if (*res)[field.From] == nil {
-			(*res)[field.From] = map[string][]string{}
+		if (*res)[getFromKey(field)] == nil {
+			(*res)[getFromKey(field)] = map[string][]string{}
 		}
 		for key, val := range valueMap {
 			resKey := key
 			// avoid article key to be duplicate
-			if vari.Def.Type == constant.ConfigTypeArticle {
+			if vari.GlobalVars.DefData.Type == consts.DefTypeArticle {
 				resKey = resKey + "_" + field.Field
 			}
-			(*res)[field.From][resKey] = val
+			(*res)[getFromKey(field)][resKey] = val
 		}
 
 	} else if field.Config != "" { // from a config
@@ -74,7 +75,7 @@ func loadResForFieldRecursive(field *model.DefField, res *map[string]map[string]
 	}
 }
 
-func getResValue(resFile, resType, sheet string, field *model.DefField) (map[string][]string, string) {
+func getResValue(resFile, resType, sheet string, field *domain.DefField) (map[string][]string, string) {
 	resName := ""
 	groupedValues := map[string][]string{}
 
@@ -87,8 +88,8 @@ func getResValue(resFile, resType, sheet string, field *model.DefField) (map[str
 	return groupedValues, resName
 }
 
-func getResFromExcel(resFile, sheet string, field *model.DefField) map[string][]string { // , string) {
-	valueMap := generateFieldValuesFromExcel(resFile, sheet, field, vari.Total)
+func getResFromExcel(resFile, sheet string, field *domain.DefField) map[string][]string { // , string) {
+	valueMap := generateFieldValuesFromExcel(resFile, sheet, field, vari.GlobalVars.Total)
 
 	return valueMap
 }
@@ -100,31 +101,34 @@ func getResFromYaml(resFile string) (valueMap map[string][]string) { // , resNam
 	}
 
 	yamlContent, err := ioutil.ReadFile(resFile)
-	yamlContent = stringUtils.ReplaceSpecialChars(yamlContent)
+	yamlContent = helper.ReplaceSpecialChars(yamlContent)
 
 	if err != nil {
 		logUtils.PrintTo(i118Utils.I118Prt.Sprintf("fail_to_read_file", resFile))
 		return
 	}
 
-	insts := model.ResInstances{}
+	insts := domain.ResInstances{}
 	err = yaml.Unmarshal(yamlContent, &insts)
 	if err == nil && insts.Instances != nil && len(insts.Instances) > 0 { // instances
 		insts.FileDir = fileUtils.GetAbsDir(resFile)
 		valueMap = getResFromInstances(insts)
 		//resName = insts.Field
+
 	} else {
-		ranges := model.ResRanges{}
+		ranges := domain.ResRanges{}
 		err = yaml.Unmarshal(yamlContent, &ranges)
 		if err == nil && ranges.Ranges != nil && len(ranges.Ranges) > 0 { // ranges
 			valueMap = getResFromRanges(ranges)
 			//resName = ranges.Field
+
 		} else {
-			configRes := model.DefField{}
+			configRes := domain.DefField{}
 			err = yaml.Unmarshal(yamlContent, &configRes)
 			if err == nil { // config
 				valueMap = getResForConfig(configRes)
 				//resName = configRes.Field
+
 			}
 		}
 	}
@@ -135,7 +139,7 @@ func getResFromYaml(resFile string) (valueMap map[string][]string) { // , resNam
 	return
 }
 
-func getResFromInstances(insts model.ResInstances) (groupedValue map[string][]string) {
+func getResFromInstances(insts domain.ResInstances) (groupedValue map[string][]string) {
 	groupedValue = map[string][]string{}
 
 	for _, inst := range insts.Instances {
@@ -146,25 +150,24 @@ func getResFromInstances(insts model.ResInstances) (groupedValue map[string][]st
 		// gen values
 		fieldFromInst := convertInstantToField(insts, inst)
 		group := inst.Instance
-		groupedValue[group] = GenerateForFieldRecursive(&fieldFromInst, false)
+		groupedValue[group] = GenerateForFieldRecursive(&fieldFromInst, false, vari.GlobalVars.Total)
 	}
 
 	return groupedValue
 }
 
-func getResFromRanges(ranges model.ResRanges) map[string][]string {
+func getResFromRanges(ranges domain.ResRanges) map[string][]string {
 	groupedValue := map[string][]string{}
 
 	for group, expression := range ranges.Ranges {
 		field := convertRangesToField(ranges, expression)
-
-		groupedValue[group] = GenerateForFieldRecursive(&field, false)
+		groupedValue[group] = GenerateForFieldRecursive(&field, false, vari.GlobalVars.Total)
 	}
 
 	return groupedValue
 }
 
-func prepareNestedInstanceRes(insts model.ResInstances, inst model.ResInstancesItem, instField model.DefField) {
+func prepareNestedInstanceRes(insts domain.ResInstances, inst domain.ResInstancesItem, instField domain.DefField) {
 	// set "from" val from parent if needed
 	if instField.From == "" {
 		if insts.From != "" {
@@ -177,7 +180,7 @@ func prepareNestedInstanceRes(insts model.ResInstances, inst model.ResInstancesI
 	instField.FileDir = insts.FileDir
 
 	if instField.Use != "" { // refer to another instances or ranges
-		if vari.Res[instField.From] == nil {
+		if vari.Res[getFromKey(&instField)] == nil {
 			referencedRanges, referencedInstants := getReferencedRangeOrInstant(instField)
 			groupedValueReferenced := map[string][]string{}
 
@@ -185,7 +188,7 @@ func prepareNestedInstanceRes(insts model.ResInstances, inst model.ResInstancesI
 				groupedValueReferenced = getResFromRanges(referencedRanges)
 
 			} else if len(referencedInstants.Instances) > 0 { // refer to instances
-				for _, referencedInst := range referencedInstants.Instances { // iterate items
+				for _, referencedInst := range referencedInstants.Instances { // iterate records
 					for _, referencedInstField := range referencedInst.Fields { // if item had children, iterate children
 						prepareNestedInstanceRes(referencedInstants, referencedInst, referencedInstField)
 					}
@@ -194,24 +197,24 @@ func prepareNestedInstanceRes(insts model.ResInstances, inst model.ResInstancesI
 
 					// gen values
 					group := referencedInst.Instance
-					groupedValueReferenced[group] = GenerateForFieldRecursive(&field, false)
+					groupedValueReferenced[group] = GenerateForFieldRecursive(&field, false, vari.GlobalVars.Total)
 				}
 			}
 
-			vari.Res[instField.From] = groupedValueReferenced
+			vari.Res[getFromKey(&instField)] = groupedValueReferenced
 		}
 	} else if instField.Select != "" { // refer to excel
 		resFile, resType, sheet := fileUtils.GetResProp(instField.From, instField.FileDir)
 		values, _ := getResValue(resFile, resType, sheet, &instField)
-		vari.Res[instField.From] = values
+		vari.Res[getFromKey(&instField)] = values
 	}
 }
 
-func getReferencedRangeOrInstant(inst model.DefField) (referencedRanges model.ResRanges, referencedInsts model.ResInstances) {
+func getReferencedRangeOrInstant(inst domain.DefField) (referencedRanges domain.ResRanges, referencedInsts domain.ResInstances) {
 	resFile, _, _ := fileUtils.GetResProp(inst.From, inst.FileDir)
 
 	yamlContent, err := ioutil.ReadFile(resFile)
-	yamlContent = stringUtils.ReplaceSpecialChars(yamlContent)
+	yamlContent = helper.ReplaceSpecialChars(yamlContent)
 	if err != nil {
 		logUtils.PrintTo(i118Utils.I118Prt.Sprintf("fail_to_read_file", resFile))
 		return
@@ -233,11 +236,11 @@ func getReferencedRangeOrInstant(inst model.DefField) (referencedRanges model.Re
 	return
 }
 
-func convertInstantToField(insts model.ResInstances, inst model.ResInstancesItem) (field model.DefField) {
+func convertInstantToField(insts domain.ResInstances, inst domain.ResInstancesItem) (field domain.DefField) {
 	//field.Field = insts.Field
 	field.From = insts.From
 
-	child := model.DefField{}
+	child := domain.DefField{}
 	child.Field = inst.Instance
 
 	// some props are from parent instances
@@ -255,18 +258,18 @@ func convertInstantToField(insts model.ResInstances, inst model.ResInstancesItem
 	return field
 }
 
-func convertRangesToField(ranges model.ResRanges, expression string) (field model.DefField) {
+func convertRangesToField(ranges domain.ResRanges, expression string) (field domain.DefField) {
 	copier.Copy(&field, ranges)
 	field.Range = expression
 
 	return field
 }
 
-func getResForConfig(configRes model.DefField) map[string][]string {
+func getResForConfig(configRes domain.DefField) map[string][]string {
 	groupedValue := map[string][]string{}
 
 	// config field is a standard field
-	groupedValue["all"] = GenerateForFieldRecursive(&configRes, false)
+	groupedValue["all"] = GenerateForFieldRecursive(&configRes, false, vari.GlobalVars.Total)
 
 	return groupedValue
 }
