@@ -2,8 +2,16 @@ package helper
 
 import (
 	"fmt"
+	"io/ioutil"
+	"os"
+	"path/filepath"
+	"regexp"
+	"sort"
+	"strings"
+
 	"github.com/360EntSecGroup-Skylar/excelize/v2"
-	constant "github.com/easysoft/zendata/internal/pkg/const"
+	consts "github.com/easysoft/zendata/internal/pkg/const"
+	"github.com/easysoft/zendata/internal/pkg/domain"
 	"github.com/easysoft/zendata/internal/pkg/model"
 	fileUtils "github.com/easysoft/zendata/pkg/utils/file"
 	i118Utils "github.com/easysoft/zendata/pkg/utils/i118"
@@ -11,40 +19,40 @@ import (
 	"github.com/easysoft/zendata/pkg/utils/vari"
 	"github.com/mattn/go-runewidth"
 	"gopkg.in/yaml.v3"
-	"io/ioutil"
-	"path/filepath"
-	"regexp"
-	"sort"
-	"strings"
 )
 
 func ListData() {
-	res, nameWidth, titleWidth := LoadRes("work")
+	res := map[string][]domain.ResFile{}
+	GetFilesAndDirs(vari.ZdDir, consts.ResDirUsers, &res)
+
+	res, nameWidth, titleWidth := LoadRes(res)
 	PrintRes(res, nameWidth, titleWidth)
 }
 
-func ListRes() {
-	res, nameWidth, titleWidth := LoadRes("zd")
+func ListRes() (ret map[string][]domain.ResFile) {
+	res, nameWidth, titleWidth := GetRes()
 	PrintRes(res, nameWidth, titleWidth)
+
+	return
 }
 
-func LoadRes(resSrc string) (res map[string][]model.ResFile, nameWidth, titleWidth int) {
-	res = map[string][]model.ResFile{}
+func GetRes() (ret map[string][]domain.ResFile, nameWidth, titleWidth int) {
+	ret = map[string][]domain.ResFile{}
 
-	if vari.WorkDir == vari.ZdPath {
-		resSrc = "zd"
+	for _, key := range consts.ResKeys {
+		GetFilesAndDirs(key, key, &ret)
 	}
 
-	if resSrc == "work" {
-		GetFilesAndDirs(vari.WorkDir, constant.ResDirUsers, &res)
-	} else {
-		for _, key := range constant.ResKeys {
-			GetFilesAndDirs(key, key, &res)
-		}
-	}
+	ret, nameWidth, titleWidth = LoadRes(ret)
 
-	for _, key := range constant.ResKeys {
-		arr := make([]model.ResFile, 0)
+	return
+}
+
+func LoadRes(res map[string][]domain.ResFile) (ret map[string][]domain.ResFile, nameWidth, titleWidth int) {
+	ret = map[string][]domain.ResFile{}
+
+	for _, key := range consts.ResKeys {
+		arr := make([]domain.ResFile, 0)
 
 		for _, item := range res[key] {
 			pth := item.Path
@@ -53,9 +61,9 @@ func LoadRes(resSrc string) (res map[string][]model.ResFile, nameWidth, titleWid
 			isArticleFiles := false
 			var title, desc, tp string
 
-			if key == constant.ResDirData { // data dir contains excel
+			if key == consts.ResDirData { // data dir contains excel
 				title, desc, tp = ReadExcelInfo(pth)
-			} else if key == constant.ResDirYaml || key == constant.ResDirUsers {
+			} else if key == consts.ResDirYaml || key == consts.ResDirUsers {
 				isArticleFiles, _ = regexp.MatchString("yaml.article", pth)
 
 				if fileExt == ".txt" { // ignore packaged article text file
@@ -75,7 +83,7 @@ func LoadRes(resSrc string) (res map[string][]model.ResFile, nameWidth, titleWid
 				nameWidth = lent
 			}
 
-			if key == constant.ResDirData {
+			if key == consts.ResDirData {
 				sheets := strings.Split(title, "|")
 				for _, sheet := range sheets {
 					lent2 := runewidth.StringWidth(sheet)
@@ -95,18 +103,18 @@ func LoadRes(resSrc string) (res map[string][]model.ResFile, nameWidth, titleWid
 			}
 		}
 
-		res[key] = SortByName(arr)
+		ret[key] = SortByName(arr)
 	}
 
 	return
 }
 
-func PrintRes(res map[string][]model.ResFile, nameWidth, titleWidth int) {
+func PrintRes(res map[string][]domain.ResFile, nameWidth, titleWidth int) {
 	dataMsg := ""
 	yamlMsg := ""
 	usersMsg := ""
 	idx := 0
-	for _, key := range constant.ResKeys {
+	for _, key := range consts.ResKeys {
 		arr := res[key]
 
 		for _, item := range arr {
@@ -125,11 +133,11 @@ func PrintRes(res map[string][]model.ResFile, nameWidth, titleWidth int) {
 				title = title + strings.Repeat(" ", titleWidth-runewidth.StringWidth(title))
 				msg := fmt.Sprintf("%s  %s  %s\n", name, title, desc)
 
-				if key == constant.ResDirData {
+				if key == consts.ResDirData {
 					dataMsg = dataMsg + msg
-				} else if key == constant.ResDirYaml {
+				} else if key == consts.ResDirYaml {
 					yamlMsg = yamlMsg + msg
-				} else if key == constant.ResDirUsers {
+				} else if key == consts.ResDirUsers {
 					usersMsg = usersMsg + msg
 				}
 
@@ -143,9 +151,9 @@ func PrintRes(res map[string][]model.ResFile, nameWidth, titleWidth int) {
 	logUtils.PrintTo(dataMsg + "\n" + yamlMsg + "\n" + usersMsg)
 }
 
-func GetFilesAndDirs(pth, typ string, res *map[string][]model.ResFile) {
+func GetFilesAndDirs(pth, typ string, res *map[string][]domain.ResFile) {
 	if !fileUtils.IsAbsPath(pth) {
-		pth = vari.ZdPath + pth
+		pth = vari.ZdDir + pth
 	}
 
 	dir, err := ioutil.ReadDir(pth)
@@ -163,14 +171,14 @@ func GetFilesAndDirs(pth, typ string, res *map[string][]model.ResFile) {
 				continue
 			}
 
-			file := model.ResFile{Path: filepath.Join(pth, name), UpdatedAt: fi.ModTime()}
+			file := domain.ResFile{Path: filepath.Join(pth, name), UpdatedAt: fi.ModTime()}
 			(*res)[typ] = append((*res)[typ], file)
 		}
 	}
 }
 
 func ReadYamlInfo(path string) (title, desc, resType string) {
-	info := model.DefInfo{}
+	info := domain.DefInfo{}
 
 	if strings.Index(path, "apache") > -1 {
 		logUtils.PrintTo("")
@@ -208,27 +216,27 @@ func ReadExcelInfo(path string) (title, desc, resType string) {
 	}
 
 	desc = i118Utils.I118Prt.Sprintf("excel_data")
-	resType = constant.ResTypeExcel
+	resType = consts.ResTypeExcel
 	return
 }
 
 func ReadTextInfo(path, key string) (title, desc, resType string) {
-	title = PathToName(path, key, constant.ResTypeText)
+	title = PathToName(path, key, consts.ResTypeText)
 	desc = i118Utils.I118Prt.Sprintf("text_data")
-	resType = constant.ResTypeText
+	resType = consts.ResTypeText
 	return
 }
 
 func PathToName(path, key, tp string) string {
-	isWorkData := strings.Index(path, vari.WorkDir) > -1
+	isWorkData := strings.Index(path, vari.ZdDir) > -1
 	if isWorkData { // user data in workdir
-		path = strings.Replace(path, vari.WorkDir, "", 1)
+		path = strings.Replace(path, vari.ZdDir, "", 1)
 	}
 
-	nameSep := constant.PthSep
-	if tp != constant.ResTypeText && tp != constant.ResTypeYaml && tp != constant.ResTypeConfig {
+	nameSep := consts.PthSep
+	if tp != consts.ResTypeText && tp != consts.ResTypeYaml && tp != consts.ResTypeConfig {
 		nameSep = "."
-		path = strings.ReplaceAll(path, constant.PthSep, nameSep)
+		path = strings.ReplaceAll(path, consts.PthSep, nameSep)
 		path = path[strings.Index(path, nameSep)+len(nameSep):]
 	}
 	if isWorkData {
@@ -237,7 +245,7 @@ func PathToName(path, key, tp string) string {
 
 	sep := nameSep + key + nameSep
 	name := path[strings.Index(path, sep)+len(sep):]
-	if key == constant.ResDirData { // remove .xlsx postfix for excel data
+	if key == consts.ResDirData { // remove .xlsx postfix for excel data
 		name = name[:strings.LastIndex(name, nameSep)]
 	}
 
@@ -248,7 +256,7 @@ func removeDirPrefix(name, seq string) (ret string) {
 	return
 }
 
-func SortByName(arr []model.ResFile) []model.ResFile {
+func SortByName(arr []domain.ResFile) []domain.ResFile {
 	sort.Slice(arr, func(i, j int) bool {
 		flag := false
 		if arr[i].ReferName > (arr[j].ReferName) {
@@ -259,16 +267,35 @@ func SortByName(arr []model.ResFile) []model.ResFile {
 	return arr
 }
 
-func GetYamlResType(def model.DefInfo) string {
+func GetYamlResType(def domain.DefInfo) string {
 	if def.Ranges != nil {
-		return constant.ResTypeRanges
+		return consts.ResTypeRanges
 	} else if def.Instances != nil {
-		return constant.ResTypeInstances
+		return consts.ResTypeInstances
 	} else if def.Fields != nil {
-		return constant.ResTypeYaml
+		return consts.ResTypeYaml
 	} else {
-		return constant.ResTypeConfig
+		return consts.ResTypeConfig
 	}
 
 	return ""
+}
+
+func GetDefFromYamlFile(path string) (po *model.ZdDef, content string, err error) {
+	po = &model.ZdDef{}
+	contentByte, _ := os.ReadFile(path)
+	content = string(contentByte)
+
+	yamlContent := ReplaceSpecialChars(contentByte)
+	err = yaml.Unmarshal(yamlContent, po)
+
+	return
+}
+
+func GetDefFromYamlString(content string) (po *model.ZdDef, err error) {
+	po = &model.ZdDef{}
+	yamlContent := ReplaceSpecialChars([]byte(content))
+	err = yaml.Unmarshal(yamlContent, po)
+
+	return
 }
