@@ -3,5 +3,5 @@ package model
 type Config struct {
 	Version  float64
 	Language string
-	//Server string
+	//DBType string
 }
