@@ -1,7 +1,8 @@
 package serverRepo
 
 import (
-	constant "github.com/easysoft/zendata/internal/pkg/const"
+	consts "github.com/easysoft/zendata/internal/pkg/const"
+	"github.com/easysoft/zendata/internal/pkg/domain"
 	"github.com/easysoft/zendata/internal/pkg/model"
 	"gorm.io/gorm"
 )
@@ -16,12 +17,12 @@ func (r *DefRepo) ListAll() (models []*model.ZdDef) {
 }
 
 func (r *DefRepo) List(keywords string, page int) (models []*model.ZdDef, total int, err error) {
-	query := r.DB.Select("id,title,folder,path,referName").Order("id ASC")
+	query := r.DB.Select("id,title,folder,path,referName").Order("id ASC").Where("isMock = false")
 	if keywords != "" {
 		query = query.Where("title LIKE ?", "%"+keywords+"%")
 	}
 	if page > 0 {
-		query = query.Offset((page - 1) * constant.PageSize).Limit(constant.PageSize)
+		query = query.Offset((page - 1) * consts.PageSize).Limit(consts.PageSize)
 	}
 
 	err = query.Find(&models).Error
@@ -66,11 +67,11 @@ func (r *DefRepo) UpdateYaml(po model.ZdDef) (err error) {
 	return
 }
 
-func (r *DefRepo) GenDef(def model.ZdDef, data *model.DefData) {
+func (r *DefRepo) GenDef(def model.ZdDef, data *domain.DefData) {
 	data.Title = def.Title
 	data.Desc = def.Desc
 	data.Type = def.Type
-	if data.Type == constant.ResTypeText {
+	if data.Type == consts.ResTypeText {
 		data.Type = ""
 	}
 }
