@@ -1,7 +1,7 @@
 package serverRepo
 
 import (
-	constant "github.com/easysoft/zendata/internal/pkg/const"
+	consts "github.com/easysoft/zendata/internal/pkg/const"
 	"github.com/easysoft/zendata/internal/pkg/model"
 	"gorm.io/gorm"
 )
@@ -21,7 +21,7 @@ func (r *ExcelRepo) List(keywords string, page int) (models []*model.ZdExcel, to
 		query = query.Where("title LIKE ?", "%"+keywords+"%")
 	}
 	if page > 0 {
-		query = query.Offset((page - 1) * constant.PageSize).Limit(constant.PageSize)
+		query = query.Offset((page - 1) * consts.PageSize).Limit(consts.PageSize)
 	}
 
 	err = query.Find(&models).Error
