@@ -1,7 +1,8 @@
 package serverRepo
 
 import (
-	constant "github.com/easysoft/zendata/internal/pkg/const"
+	consts "github.com/easysoft/zendata/internal/pkg/const"
+	"github.com/easysoft/zendata/internal/pkg/domain"
 	"github.com/easysoft/zendata/internal/pkg/model"
 	"gorm.io/gorm"
 )
@@ -21,7 +22,7 @@ func (r *InstancesRepo) List(keywords string, page int) (models []*model.ZdInsta
 		query = query.Where("title LIKE ?", "%"+keywords+"%")
 	}
 	if page > 0 {
-		query = query.Offset((page - 1) * constant.PageSize).Limit(constant.PageSize)
+		query = query.Offset((page - 1) * consts.PageSize).Limit(consts.PageSize)
 	}
 
 	err = query.Find(&models).Error
@@ -140,7 +141,7 @@ func (r *InstancesRepo) haveChild(Data []*model.ZdInstancesItem,
 	return
 }
 
-func (r *InstancesRepo) GenInst(po model.ZdInstances, res *model.ResInstances) {
+func (r *InstancesRepo) GenInst(po model.ZdInstances, res *domain.ResInstances) {
 	res.Title = po.Title
 	res.Desc = po.Desc
 }
