@@ -1,9 +1,10 @@
 package serverService
 
 import (
+	"github.com/easysoft/zendata/internal/pkg/domain"
 	"strings"
 
-	constant "github.com/easysoft/zendata/internal/pkg/const"
+	consts "github.com/easysoft/zendata/internal/pkg/const"
 	"github.com/easysoft/zendata/internal/pkg/helper"
 	"github.com/easysoft/zendata/internal/pkg/model"
 	serverRepo "github.com/easysoft/zendata/internal/server/repo"
@@ -23,18 +24,18 @@ func (s *ExcelService) List(keywords string, page int) (list []*model.ZdExcel, t
 	return
 }
 
-func (s *ExcelService) Get(id int) (excel model.ZdExcel, dirs []model.Dir) {
+func (s *ExcelService) Get(id int) (excel model.ZdExcel, dirs []domain.Dir) {
 	excel, _ = s.ExcelRepo.Get(uint(id))
 
-	serverUtils.GetDirs(constant.ResDirData, &dirs)
+	serverUtils.GetDirs(consts.ResDirData, &dirs)
 
 	return
 }
 
 func (s *ExcelService) Save(excel *model.ZdExcel) (err error) {
 	excel.Folder = serverUtils.DealWithPathSepRight(excel.Folder)
-	excel.Path = vari.ZdPath + excel.Folder + serverUtils.AddExt(excel.FileName, ".xlsx")
-	excel.ReferName = helper.PathToName(excel.Path, constant.ResDirData, constant.ResTypeExcel)
+	excel.Path = vari.ZdDir + excel.Folder + serverUtils.AddExt(excel.FileName, ".xlsx")
+	excel.ReferName = helper.PathToName(excel.Path, consts.ResDirData, consts.ResTypeExcel)
 
 	if excel.ID == 0 {
 		// excel should not be create on webpage
@@ -71,7 +72,7 @@ func (s *ExcelService) Remove(id int) (err error) {
 	return
 }
 
-func (s *ExcelService) Sync(files []model.ResFile) (err error) {
+func (s *ExcelService) Sync(files []domain.ResFile) (err error) {
 	list := s.ExcelRepo.ListAll()
 
 	mp := map[string]*model.ZdExcel{}
@@ -81,7 +82,7 @@ func (s *ExcelService) Sync(files []model.ResFile) (err error) {
 
 	for _, fi := range files {
 		_, found := mp[fi.Path]
-		//logUtils.PrintTo(fi.UpdatedAt.String() + ", " + mp[fi.Path].UpdatedAt.String())
+		//logUtils.PrintTo(fi.UpdatedAt.OpenApiDataTypeString() + ", " + mp[fi.Path].UpdatedAt.OpenApiDataTypeString())
 		if !found { // no record
 			s.SyncToDB(fi)
 		} else if fi.UpdatedAt.Unix() > mp[fi.Path].UpdatedAt.Unix() { // db is old
@@ -94,13 +95,13 @@ func (s *ExcelService) Sync(files []model.ResFile) (err error) {
 
 	return
 }
-func (s *ExcelService) SyncToDB(file model.ResFile) (err error) {
+func (s *ExcelService) SyncToDB(file domain.ResFile) (err error) {
 	excel := model.ZdExcel{
 		Title:     file.Title,
 		Sheet:     file.Title,
 		Path:      file.Path,
 		Folder:    serverUtils.GetRelativePath(file.Path),
-		ReferName: helper.PathToName(file.Path, constant.ResDirData, constant.ResTypeExcel),
+		ReferName: helper.PathToName(file.Path, consts.ResDirData, consts.ResTypeExcel),
 		FileName:  fileUtils.GetFileName(file.Path),
 	}
 	s.ExcelRepo.Create(&excel)
