@@ -3,7 +3,8 @@ package serverService
 import (
 	"fmt"
 	"github.com/easysoft/zendata/internal/pkg/action"
-	constant "github.com/easysoft/zendata/internal/pkg/const"
+	consts "github.com/easysoft/zendata/internal/pkg/const"
+	"github.com/easysoft/zendata/internal/pkg/domain"
 	"github.com/easysoft/zendata/internal/pkg/model"
 	"github.com/easysoft/zendata/internal/server/repo"
 	fileUtils "github.com/easysoft/zendata/pkg/utils/file"
@@ -22,7 +23,8 @@ type PreviewService struct {
 func (s *PreviewService) PreviewDefData(defId uint) (data string) {
 	def, _ := s.DefRepo.Get(defId)
 
-	lines := action.Generate([]string{def.Path}, "", constant.FormatData, "")
+	vari.GlobalVars.Total = 10
+	lines := action.Generate([]string{def.Path}, "", consts.FormatData, "")
 	data = s.linesToStr(lines)
 
 	return
@@ -30,9 +32,9 @@ func (s *PreviewService) PreviewDefData(defId uint) (data string) {
 func (s *PreviewService) PreviewFieldData(fieldId uint, fieldType string) (data string) {
 	var field model.ZdField
 
-	if fieldType == constant.ResTypeDef {
+	if fieldType == consts.ResTypeDef {
 		field, _ = s.FieldRepo.Get(fieldId)
-	} else if fieldType == constant.ResTypeInstances {
+	} else if fieldType == consts.ResTypeInstances {
 		instItem, _ := s.InstancesRepo.GetItem(fieldId)
 		field.From = instItem.From
 		copier.Copy(&field, instItem)
@@ -40,20 +42,20 @@ func (s *PreviewService) PreviewFieldData(fieldId uint, fieldType string) (data 
 
 	ref := model.ZdRefer{}
 	if !field.IsRange {
-		ref, _ = s.ReferRepo.GetByOwnerId(field.ID)
+		ref, _ = s.ReferRepo.GetByOwnerIdAndType(field.ID, fieldType)
 	}
 
-	fld := model.DefField{}
+	fld := domain.DefField{}
 	genFieldFromZdField(field, ref, &fld)
 
-	def := model.DefData{}
+	def := domain.DefData{}
 	def.Fields = append(def.Fields, fld)
 	defContent, _ := yaml.Marshal(def)
 
-	configFile := vari.ZdPath + "tmp" + constant.PthSep + ".temp.yaml"
+	configFile := vari.ZdDir + "tmp" + consts.PthSep + ".temp.yaml"
 	fileUtils.WriteFile(configFile, string(defContent))
 
-	lines := action.Generate([]string{configFile}, field.Field, constant.FormatData, "")
+	lines := action.Generate([]string{configFile}, field.Field, consts.FormatData, "")
 	data = s.linesToStr(lines)
 
 	return
