@@ -1,9 +1,10 @@
 package serverService
 
 import (
+	"github.com/easysoft/zendata/internal/pkg/domain"
 	"strings"
 
-	constant "github.com/easysoft/zendata/internal/pkg/const"
+	consts "github.com/easysoft/zendata/internal/pkg/const"
 	"github.com/easysoft/zendata/internal/pkg/helper"
 	"github.com/easysoft/zendata/internal/pkg/model"
 	serverRepo "github.com/easysoft/zendata/internal/server/repo"
@@ -24,18 +25,18 @@ func (s *TextService) List(keywords string, page int) (list []*model.ZdText, tot
 	return
 }
 
-func (s *TextService) Get(id int) (text model.ZdText, dirs []model.Dir) {
+func (s *TextService) Get(id int) (text model.ZdText, dirs []domain.Dir) {
 	text, _ = s.TextRepo.Get(uint(id))
 
-	serverUtils.GetDirs(constant.ResDirYaml, &dirs)
+	serverUtils.GetDirs(consts.ResDirYaml, &dirs)
 
 	return
 }
 
 func (s *TextService) Save(text *model.ZdText) (err error) {
 	text.Folder = serverUtils.DealWithPathSepRight(text.Folder)
-	text.Path = vari.ZdPath + text.Folder + serverUtils.AddExt(text.FileName, ".txt")
-	text.ReferName = helper.PathToName(text.Path, constant.ResDirYaml, constant.ResTypeText)
+	text.Path = vari.ZdDir + text.Folder + serverUtils.AddExt(text.FileName, ".txt")
+	text.ReferName = helper.PathToName(text.Path, consts.ResDirYaml, consts.ResTypeText)
 
 	if text.ID == 0 {
 		err = s.Create(text)
@@ -78,7 +79,7 @@ func (s *TextService) Remove(id int) (err error) {
 	return
 }
 
-func (s *TextService) Sync(files []model.ResFile) (err error) {
+func (s *TextService) Sync(files []domain.ResFile) (err error) {
 	list := s.TextRepo.ListAll()
 
 	mp := map[string]*model.ZdText{}
@@ -88,7 +89,7 @@ func (s *TextService) Sync(files []model.ResFile) (err error) {
 
 	for _, fi := range files {
 		_, found := mp[fi.Path]
-		//logUtils.PrintTo(fi.UpdatedAt.String() + ", " + mp[fi.Path].UpdatedAt.String())
+		//logUtils.PrintTo(fi.UpdatedAt.OpenApiDataTypeString() + ", " + mp[fi.Path].UpdatedAt.OpenApiDataTypeString())
 		if !found { // no record
 			s.SyncToDB(fi)
 		} else if fi.UpdatedAt.Unix() > mp[fi.Path].UpdatedAt.Unix() { // db is old
@@ -102,17 +103,17 @@ func (s *TextService) Sync(files []model.ResFile) (err error) {
 	return
 }
 
-func (s *TextService) SyncToDB(file model.ResFile) (err error) {
+func (s *TextService) SyncToDB(file domain.ResFile) (err error) {
 	text := model.ZdText{
 		Title:    file.Title,
 		Path:     file.Path,
 		Folder:   serverUtils.GetRelativePath(file.Path),
 		FileName: fileUtils.GetFileName(file.Path),
 	}
-	if strings.Index(text.Path, constant.ResDirYaml) > -1 {
-		text.ReferName = helper.PathToName(text.Path, constant.ResDirYaml, constant.ResTypeText)
+	if strings.Index(text.Path, consts.ResDirYaml) > -1 {
+		text.ReferName = helper.PathToName(text.Path, consts.ResDirYaml, consts.ResTypeText)
 	} else {
-		text.ReferName = helper.PathToName(text.Path, constant.ResDirUsers, constant.ResTypeText)
+		text.ReferName = helper.PathToName(text.Path, consts.ResDirUsers, consts.ResTypeText)
 	}
 	text.Content = fileUtils.ReadFile(file.Path)
 
