@@ -2,7 +2,8 @@ package serverUtils
 
 import (
 	"encoding/json"
-	constant "github.com/easysoft/zendata/internal/pkg/const"
+	consts "github.com/easysoft/zendata/internal/pkg/const"
+	"github.com/easysoft/zendata/internal/pkg/domain"
 	"github.com/easysoft/zendata/internal/pkg/model"
 	fileUtils "github.com/easysoft/zendata/pkg/utils/file"
 	"github.com/easysoft/zendata/pkg/utils/vari"
@@ -89,9 +90,9 @@ func ConvertParams(data interface{}) (mp map[string]string) {
 }
 
 func GetRelativePath(pth string) string {
-	idx := strings.LastIndex(pth, constant.PthSep)
+	idx := strings.LastIndex(pth, consts.PthSep)
 	folder := pth[:idx+1]
-	folder = strings.Replace(folder, vari.ZdPath, "", 1)
+	folder = strings.Replace(folder, vari.ZdDir, "", 1)
 
 	return folder
 }
@@ -111,14 +112,14 @@ func AddExt(pth, ext string) string {
 	return pth
 }
 
-func GetDirs(dir string, dirs *[]model.Dir) {
+func GetDirs(dir string, dirs *[]domain.Dir) {
 	isArticleFiles, _ := regexp.MatchString("yaml.article", dir)
 	if isArticleFiles {
 		return
 	}
 
 	folder := fileUtils.AddSepIfNeeded(dir)
-	*dirs = append(*dirs, model.Dir{Name: folder})
+	*dirs = append(*dirs, domain.Dir{Name: folder})
 
 	files, _ := ioutil.ReadDir(folder)
 	for _, fi := range files {
