@@ -1,7 +1,8 @@
 package commonUtils
 
 import (
-	"github.com/easysoft/zendata/internal/pkg/const"
+	"fmt"
+	consts "github.com/easysoft/zendata/internal/pkg/const"
 	"github.com/easysoft/zendata/internal/pkg/model"
 	stringUtils "github.com/easysoft/zendata/pkg/utils/string"
 	"github.com/emirpasic/gods/maps"
@@ -12,6 +13,7 @@ import (
 	"reflect"
 	"regexp"
 	"runtime"
+	"strconv"
 	"strings"
 	"time"
 )
@@ -33,9 +35,9 @@ func RemoveBlankLine(str string) string {
 
 func BoolToPass(b bool) string {
 	if b {
-		return constant.PASS.String()
+		return consts.PASS.String()
 	} else {
-		return constant.FAIL.String()
+		return consts.FAIL.String()
 	}
 }
 
@@ -61,7 +63,7 @@ func IsMac() bool {
 func IsRelease() bool {
 	arg1 := strings.ToLower(os.Args[0])
 	name := filepath.Base(arg1)
-	return strings.Index(name, constant.AppName) == 0 && strings.Index(arg1, "go-build") < 0
+	return strings.Index(name, consts.AppName) == 0 && strings.Index(arg1, "go-build") < 0
 
 	//if _, err := os.Stat("res"); os.IsNotExist(err) {
 	//	return true
@@ -211,4 +213,25 @@ func RandNum64(length int64) int64 {
 
 	seedInt := rand.Int63n(length)
 	return seedInt
+}
+
+func ChanePrecision(value float64, precision int) float64 {
+	temp := fmt.Sprintf("%."+strconv.Itoa(precision)+"f", value)
+	value, _ = strconv.ParseFloat(temp, 64)
+	return value
+}
+
+func GetType(i interface{}) string {
+	switch i.(type) {
+	case int64:
+		return "int"
+	case float64:
+		return "float"
+	case byte:
+		return "byte"
+	case string:
+		return "string"
+	}
+
+	return "string"
 }
