@@ -8,19 +8,26 @@ export default api
 
 export function getWorkDir () {
   return request({
-    url: api.admin,
+    url: `/comm/getWorkDir`,
+    method: 'get',
+  })
+}
+
+export function syncData () {
+  return request({
+    url: `/comm/syncData`,
     method: 'post',
-    data: {'action': 'getWorkDir'}
   })
 }
 
 export function listDef (keywords, page) {
   return request({
-    url: api.admin,
-    method: 'post',
-    data: {'action': 'listDef', keywords: keywords, page: page}
+    url: `/defs`,
+    method: 'get',
+    params: {keywords: keywords, page: page}
   })
 }
+
 export function getDef (id) {
   const data = {'action': 'getDef', id: id}
   console.log(data)
@@ -110,14 +117,15 @@ export function moveDefField (src, dist, mode) {
 }
 
 export function previewDefData (defId) {
-  const data = {'action': 'previewDefData', id: defId}
+  const params = {defId: defId}
 
   return request({
-    url: api.admin,
-    method: 'post',
-    data: data
+    url: `/defs/previewData`,
+    method: 'get',
+    params
   })
 }
+
 export function previewFieldData (fieldId, type) {
   const data = {'action': 'previewFieldData', id: fieldId, mode: type}
   return request({
@@ -422,12 +430,3 @@ export function removeExcel (id) {
   })
 }
 
-export function syncData () {
-  const data = {'action': 'syncData', mode: ''}
-
-  return request({
-    url: api.admin,
-    method: 'post',
-    data: data
-  })
-}
