@@ -36,7 +36,7 @@
 </template>
 
 <script>
-import {getExcel, saveText} from "../../../../api/manage";
+import {getExcel, saveExcel, saveText} from "../../../../api/manage";
 import {checkDirIsData} from "../../../../api/utils";
 import {
   colsFull,
@@ -47,6 +47,7 @@ import {
   wrapperColFull,
   wrapperColHalf
 } from "@/utils/const";
+import {getDir} from "@/utils/utils";
 
 export default {
   name: 'TestEdit',
@@ -81,7 +82,7 @@ export default {
         ],
       },
 
-      model: { folder: 'data/'},
+      model: { folder: getDir('data')},
       dirs: [],
       workDir: '',
     };
@@ -100,20 +101,18 @@ export default {
   methods: {
     loadData () {
       let id = this.id;
-      if (id === null) {
-        return;
-      }
-      if (id) {
-        if (typeof id === 'string') id = Number.parseInt(id);
-         getExcel(id).then(json => {
-          console.log('getText', json)
-          this.model = json.data
-          this.dirs = json.res
-          this.workDir = json.workDir
-        })
-      } else {
+      if (typeof id === 'string') id = Number.parseInt(id);
+      if (!id) {
         this.reset();
+        return
       }
+
+      getExcel(id).then(json => {
+        console.log('getText', json)
+        this.model = json.data
+        this.dirs = json.res
+        this.workDir = json.workDir
+      })
     },
     save() {
       console.log('save')
@@ -125,8 +124,8 @@ export default {
         }
 
         if (this.model.subFolder && this.model.subFolder != '') this.model.folder += this.model.subFolder
-        saveText(this.model).then(json => {
-          console.log('saveText', json)
+        saveExcel(this.model).then(json => {
+          console.log('saveExcel', json)
           if (this.afterSave) {
             this.afterSave(json);
           }
@@ -135,8 +134,8 @@ export default {
     },
     reset() {
       console.log('reset')
-      this.model = {folder: 'data/'};
-      this.$refs.editForm.reset()
+      this.$refs.editForm.resetFields()
+      this.model = {folder: getDir('data')};
     },
   }
 }
