@@ -61,6 +61,7 @@ import {
   wrapperColFull,
   wrapperColHalf
 } from "@/utils/const";
+import {getDir} from "@/utils/utils";
 
 export default {
   name: 'RangesEdit',
@@ -95,7 +96,7 @@ export default {
         ],
       },
 
-      model: {folder: 'yaml/'},
+      model: {folder: getDir('yaml')},
       dirs: [],
       workDir: '',
     };
@@ -115,20 +116,18 @@ export default {
   methods: {
     loadData () {
       let id = this.id;
-      if (id === null) {
-        return;
-      }
-      if (id) {
-        if (typeof id === 'string') id = Number.parseInt(id);
-        getRanges(id).then(json => {
-          console.log('getRanges', json)
-          this.model = json.data
-          this.dirs = json.res
-          this.workDir = json.workDir
-        })
-      } else {
+      if (typeof id === 'string') id = Number.parseInt(id);
+      if (!id) {
         this.reset();
+        return
       }
+
+      getRanges(id).then(json => {
+        console.log('getRanges', json)
+        this.model = json.data
+        this.dirs = json.res
+        this.workDir = json.workDir
+      })
     },
     save() {
       console.log('save')
@@ -150,8 +149,8 @@ export default {
     },
     reset() {
       console.log('reset')
-      this.model = {folder: 'yaml/'};
-      this.$refs.editForm.reset()
+      this.$refs.editForm.resetFields()
+      this.model = {folder: getDir('yaml')};
     },
   }
 }
